Cassandra ratings_timestamp table (saving and retrieving)

+++ api/src/main/scala/gr/ml/analytics/service/RatingService.scala
@@ -6,6 +6,6 @@ trait RatingService {
     * Create new ratings for a given user
     * @param userId id of the user who rated items
     */
-  def save(userId: Int, itemId: Int, rating: Double, timestamp: Long)
+  def save(userId: Int, itemId: Int, rating: Double, timestamp: Int)
 
 }
